+++ tests/testthat/test-attributes.R
@@ -121,6 +121,15 @@ test_that('Decimal padding works', {
 })
 
 
+test_that('Can pad with align; pad_decimal gives warning', {
+  ht <- hux(a = c("1.5", "2.5"))
+  ht2 <- ht
+  expect_silent(align(ht) <- ".")
+  expect_warning(pad_decimal(ht2) <- ".", "deprecated")
+  expect_identical(huxtable:::clean_contents(ht), huxtable:::clean_contents(ht2))
+})
+
+
 test_that('Can set attributes to NA', {
   ht <- huxtable(a = 1:3, b = 1:3)
   # expect no error